// packages/ui/src/recoil/tryout/keplrWallet/types.ts
export type AtomKeplrWalletState = {
  keplerAdderss: string | null;
  balanceGGEZKeplr: string | null;
  balanceGGEZOneKeplr: string | null;
  GGEZDenom: string | null;
  GGEZOneDenom: string | null;
  userNameKeplr: string | null;
  installKeplrError: string | null;
  addAccountKeplrError: string | null;
};